feat(util): now supports custom help sections for uncompleted path inputs
If you have a script at `$SU_SCRIPTS_PATH/command/subcommand` and the
input `command` is given to `util` it now searches for help sections at
`$SU_SCRIPTS_PATH/command/help` and prints it along with the "commands
available" section.

The "commands available" section is unchanged, it still always list the
commands available for a existing uncomplete path input.